requirements.txt
black
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
gspread
pillow
